Don't wget if file already exists

+++ download.sh
@@ -12,6 +12,10 @@ rm -f images/$gag_id.*
 okay=false
 while [[ "$okay" == "false" ]]
 do
+    if [ -e images/$gag_id.jpg ]
+    then
+        break
+    fi
     wget -q -O "images/$gag_id" "$image_url"
     desc=`file images/$gag_id`
     if [[ "$desc" =~ "GIF" ]]
